Moved things around and removed some unused autogenerated files
