+++ client/src/components/Nav/index.tsx
@@ -5,8 +5,8 @@ import Box from '@mui/material/Box';
 import MenuIcon from '@mui/icons-material/Menu';
 import * as React from 'react';
 import Button from '@mui/material/Button';
-import { StyledTitle } from './StyledTitle';
-import { AccountCircle } from "@mui/icons-material";
+import { StyledTitle } from '../shared/StyledTitle';
+import UserMenu from "./UserMenu";
 
 
 const pages = [
@@ -19,22 +19,13 @@ const pages = [
 export default function Nav() {
   const [anchorElNav, setAnchorElNav] = React.useState<null | HTMLElement>(null);
   const [auth, setAuth] = React.useState(true);
-  const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
 
   const handleOpenNavMenu = (event: React.MouseEvent<HTMLElement>) => {
     setAnchorElNav(event.currentTarget);
   };
 
-  const handleMenu = (event: React.MouseEvent<HTMLElement>) => {
-    setAnchorEl(event.currentTarget);
-  };
-
   const handleCloseNavMenu = () => {
     setAnchorElNav(null);
-  };
-
-  const handleClose = () => {
-    setAnchorEl(null);
   };
 
   return (
@@ -92,38 +83,7 @@ export default function Nav() {
             ))}
           </Box>
 
-          {auth && (
-            <div>
-              <IconButton
-                size="large"
-                aria-label="account of current user"
-                aria-controls="menu-appbar"
-                aria-haspopup="true"
-                onClick={handleMenu}
-                color="inherit"
-              >
-                <AccountCircle />
-              </IconButton>
-              <Menu
-                id="menu-appbar"
-                anchorEl={anchorEl}
-                anchorOrigin={{
-                  vertical: 'top',
-                  horizontal: 'right',
-                }}
-                keepMounted
-                transformOrigin={{
-                  vertical: 'top',
-                  horizontal: 'right',
-                }}
-                open={Boolean(anchorEl)}
-                onClose={handleClose}
-              >
-                <MenuItem onClick={handleClose}>Profile</MenuItem>
-                <MenuItem onClick={handleClose}>My account</MenuItem>
-              </Menu>
-            </div>
-          )}
+          {auth && <UserMenu/>}
         </Toolbar>
       </Container>
     </AppBar>
